--- app/src/androidTest/java/ch/epfl/sweng/SDP/Matchmaking/MatchmakerTest.java
@@ -1,30 +1,0 @@
-package ch.epfl.sweng.SDP.Matchmaking;
-
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
-
-import ch.epfl.sweng.SDP.Matchmaking.Matchmaker;
-
-public class MatchmakerTest {
-
-    @Test
-    public void matchMaker() {
-        assertTrue(true);
-    }
-
-
-    @Test
-    public void joinRoom() {
-        Matchmaker.getInstance().joinRoom();
-        assertTrue(true);
-
-    }
-
-    @Test
-    public void leaveRoom() {
-        Matchmaker.getInstance().leaveRoom("2312");
-        assertTrue(true);
-    }
-
-}
