+++ app/src/main/java/ch/epfl/sweng/SDP/matchmaking/User.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.SDP.Matchmaking;
+package ch.epfl.sweng.SDP.matchmaking;
 
 import android.annotation.TargetApi;
 import android.os.Build;
